Allow others to import and run Felix
